implement DEBUG_HTTP_TRAFFIC as a RoundTripper
This is much more flexible than the previous implementation, since we can
better control debugging now and also the location of the output. Also, people
can use the RoundTripper directly without needing to use the rest.Client.

Updates DialSocket to take a Transport now; it's a shame to change the API but
this is probably more correct than the old version.